files/requirements.txt
ccxt
pandas
pandas-ta
scipy
statsmodels
anthropic
python-telegram-bot
python-dotenv
tenacity
requests
